WIP: put a lot of Debug derives behind a target arch flag
