+++ clovers/src/materials/dielectric.rs
@@ -14,7 +14,8 @@ use crate::{
 #[cfg(not(target_arch = "spirv"))]
 use rand::Rng;
 
-#[derive(Copy, Clone, Debug)]
+#[derive(Clone, Copy)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 /// A dielectric material. This resembles glass and other transparent and reflective materials.
 pub struct Dielectric {
