+++ clovers/src/materials/isotropic.rs
@@ -11,7 +11,8 @@ use crate::{
     Float, PI,
 };
 
-#[derive(Debug, Copy, Clone, Default)]
+#[derive(Clone, Copy, Default)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 /// Isotropic material. Used in [ConstantMedium](crate::objects::constant_medium). TODO: understand this!
 pub struct Isotropic {
