+++ clovers/src/objects/flip_face.rs
@@ -10,7 +10,7 @@ use crate::{
 
 use super::Object;
 
-#[derive(Debug)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 /// FlipFaceInit structure describes the necessary data for constructing a [FlipFace]. Used with [serde] when importing [SceneFiles](crate::scenes::SceneFile).
 pub struct FlipFaceInit {
@@ -18,7 +18,8 @@ pub struct FlipFaceInit {
     pub object: Box<Object>,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Clone)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 /// An utility object that can be used to flip the face of the object. TODO: possibly deprecated?
 pub struct FlipFace {
     object: Box<Hitable>,
