+++ clovers/src/objects/quad.rs
@@ -12,8 +12,12 @@ use crate::{
 #[cfg(not(target_arch = "spirv"))]
 use rand::Rng;
 
+#[cfg(target_arch = "spirv")]
+use spirv_std::num_traits::Float as FloatTrait;
+
 /// Initialization structure for a Quad object.
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 pub struct QuadInit {
     /// Corner point
@@ -28,7 +32,8 @@ pub struct QuadInit {
 }
 
 /// Quadrilateral shape. This can be an arbitrary parallelogram, not just a rectangle.
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 pub struct Quad {
     /// Corner point
@@ -55,12 +60,22 @@ pub struct Quad {
 impl Quad {
     /// Creates a new quad
     pub fn new(q: Vec3, u: Vec3, v: Vec3, material: Material) -> Quad {
+        // TODO: better ergonomics
+        #[cfg(not(target_arch = "spirv"))]
         let n: Vec3 = u.cross(&v);
+        #[cfg(target_arch = "spirv")]
+        let n: Vec3 = u.cross(v);
         let normal: Vec3 = n.normalize();
         // TODO: what is this?
+        #[cfg(not(target_arch = "spirv"))]
         let d = -(normal.dot(&q));
+        #[cfg(target_arch = "spirv")]
+        let d = -(normal.dot(q));
         // TODO: what is this?
+        #[cfg(not(target_arch = "spirv"))]
         let w: Vec3 = n / n.dot(&n);
+        #[cfg(target_arch = "spirv")]
+        let w: Vec3 = n / n.dot(n);
         let area = n.magnitude();
         let mut aabb = AABB::new_from_coords(q, q + u + v);
         aabb.pad();
@@ -86,7 +101,11 @@ impl Quad {
         distance_max: Float,
         _rng: &mut CloversRng,
     ) -> Option<HitRecord> {
+        // TODO: better ergonomics
+        #[cfg(not(target_arch = "spirv"))]
         let denom = self.normal.dot(&ray.direction);
+        #[cfg(target_arch = "spirv")]
+        let denom = self.normal.dot(ray.direction);
 
         // No hit if the ray is parallel to the plane.
         if denom.abs() < EPSILON_RECT_THICKNESS {
@@ -94,7 +113,11 @@ impl Quad {
         }
 
         // Return false if the hit point parameter t is outside the ray interval
+        // TODO: better ergonomics
+        #[cfg(not(target_arch = "spirv"))]
         let t = (-self.d - self.normal.dot(&ray.origin)) / denom;
+        #[cfg(target_arch = "spirv")]
+        let t = (-self.d - self.normal.dot(ray.origin)) / denom;
         if t < distance_min || t > distance_max {
             return None;
         }
@@ -102,8 +125,15 @@ impl Quad {
         // Determine the hit point lies within the planar shape using its plane coordinates.
         let intersection: Vec3 = ray.evaluate(t);
         let planar_hitpt_vector: Vec3 = intersection - self.q;
+        // TODO: better ergonomics
+        #[cfg(not(target_arch = "spirv"))]
         let alpha: Float = self.w.dot(&planar_hitpt_vector.cross(&self.v));
+        #[cfg(not(target_arch = "spirv"))]
         let beta: Float = self.w.dot(&self.u.cross(&planar_hitpt_vector));
+        #[cfg(target_arch = "spirv")]
+        let alpha: Float = self.w.dot(planar_hitpt_vector.cross(self.v));
+        #[cfg(target_arch = "spirv")]
+        let beta: Float = self.w.dot(self.u.cross(planar_hitpt_vector));
 
         // Do we hit a coordinate within the surface of the plane?
         if !hit_ab(alpha, beta) {
