+++ clovers/src/scenes.rs
@@ -9,10 +9,13 @@ use crate::{
     objects::Object,
     Float, Vec,
 };
+#[cfg(feature = "rand-crate")]
+#[cfg(not(target_arch = "spirv"))]
 use rand::SeedableRng;
+#[cfg(feature = "traces")]
 use tracing::info;
 
-#[derive(Debug)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 /// A representation of the scene that is being rendered.
 pub struct Scene {
     /// Bounding-volume hierarchy of [Hitable] objects in the scene. This could, as currently written, be any [Hitable] - in practice, we place the root of the [BVHNode](crate::bvhnode::BVHNode) tree here.
@@ -47,7 +50,7 @@ impl Scene {
 }
 
 // TODO: better naming
-#[derive(Debug)]
+#[cfg_attr(not(target_arch = "spirv"), derive(Debug))]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
 /// A serialized representation of a [Scene].
 pub struct SceneFile {
@@ -77,6 +80,7 @@ pub fn initialize(scene_file: SceneFile, width: u32, height: u32) -> Scene {
         time_1,
     );
 
+    #[cfg(feature = "traces")]
     info!("Creating a flattened list from the objects");
     let hitables = objects_to_flat_hitablelist(scene_file.objects);
